+++ etc/setup.sh
@@ -1,34 +1,65 @@
 #!/bin/bash
 # install
-mkdir ~/etc/
-FILE=~/etc/var.sh
-if [ ! -f $FILE ];
+
+can_run=True
+
+# Test python version
+version=$(python -V 2>&1 | grep -Po '(?<=Python )(.+)')
+if [[ -z "$version" ]]
 then
-    >$FILE
-    echo "PythonFolder=$(dirname "$(which python)")" >> $FILE
-    echo "SitePackageFolder=$(python -c 'import site; print(site.getsitepackages()[0])')" >> $FILE 
-    
-    # Loading pycurl
-    pip uninstall pycurl
-    export PYCURL_SSL_LIBRARY=nss
-    pip install --compile --install-option="--with-nss" --no-cache-dir pycurl
+    echo "No Python!"
+    can_run=False
 fi
 
-source ~/etc/var.sh
-export PATH=$PythonFolder:$PATH
-export PYTHONPATH=$PythonFolder
-export PATH=$HOME/.local/bin:$PATH
-export DBS3_CLIENT=$SitePackageFolder/DbsClient/
-export LD_LIBRARY_PATH=$(which curl)
-export X509_USER_CERT=$HOME/.globus/usercert.pem
-export X509_USER_KEY=$HOME/.globus/userkey.pem
-
-# test
-if [ "$1" = "test" ];
+parsedVersion=$(echo "${version//./}")
+if [[ "$parsedVersion" -lt "360" && "$parsedVersion" -gt "380" && $can_run=True ]]
 then
-    export DBS_WRITER_URL=https://cmsweb-testbed.cern.ch/dbs/int/global/DBSWriter/
-    export DBS_READER_URL=https://cmsweb-prod.cern.ch/dbs/prod/global/DBSReader/
+    echo "Valid version"
+else
+    echo "Invalid version"
+    can_run=False
 fi
 
+# Create the directory etc if needed
+if [ ! -d /etc/ ];
+then
+    mkdir /etc/
+fi
 
-echo $DBS3_CLIENT
+if [ $can_run=True ]
+then
+    FILE=~/etc/var.sh
+    if [ ! -f $FILE ];
+    then
+        >$FILE
+        echo "PythonFolder=$(dirname "$(which python)")" >> $FILE
+        echo "SitePackageFolder=$(python -c 'import site; print(site.getsitepackages()[0])')" >> $FILE
+
+        # Loading pycurl
+        pip uninstall pycurl
+         export PYCURL_SSL_LIBRARY=nss
+        pip install --compile --install-option="--with-nss" --no-cache-dir pycurl
+    fi
+
+    source ~/etc/var.sh
+    export PATH=$PythonFolder:$PATH
+    export PYTHONPATH=$PythonFolder
+    export PATH=$HOME/.local/bin:$PATH
+    export DBS3_CLIENT=$SitePackageFolder/DbsClient/
+    export LD_LIBRARY_PATH=/usr/bin/
+    export X509_USER_CERT=$HOME/.globus/usercert.pem
+    export X509_USER_KEY=$HOME/.globus/userkey.pem
+
+    # test
+    if [ "$1" = "test" ];
+    then
+        export DBS_WRITER_URL=https://cmsweb-testbed.cern.ch/dbs/int/global/DBSWriter/
+        export DBS_READER_URL=https://cmsweb-prod.cern.ch/dbs/prod/global/DBSReader/
+        
+        echo $DBS_WRITER_URL
+        echo $DBS_READER_URL
+    fi
+
+
+    echo $DBS3_CLIENT
+fi
